GG-36898 fixed tests
