+++ modules/clients/src/test/java/org/apache/ignite/jdbc/thin/JdbcAbstractBulkLoadSelfTest.java
@@ -473,71 +473,6 @@ public abstract class JdbcAbstractBulkLoadSelfTest extends JdbcThinAbstractDmlSt
     }
 
     /**
-     * Verifies that no error is reported and characters are converted improperly when we import
-     * UTF-8 as windows-1251.
-     *
-     * @throws SQLException If failed.
-     */
-    @Test
-    public void testWrongCharset_Utf8AsWin1251() throws SQLException {
-        checkBulkLoadWithWrongCharset(BULKLOAD_UTF8_CSV_FILE, "UTF-8", "windows-1251");
-    }
-
-    /**
-     * Verifies that no error is reported and characters are converted improperly when we import
-     * windows-1251 as UTF-8.
-     *
-     * @throws SQLException If failed.
-     */
-    @Test
-    public void testWrongCharset_Win1251AsUtf8() throws SQLException {
-        checkBulkLoadWithWrongCharset(BULKLOAD_CP1251_CSV_FILE, "windows-1251", "UTF-8");
-    }
-
-    /**
-     * Verifies that no error is reported and characters are converted improperly when we import
-     * UTF-8 as ASCII.
-     *
-     * @throws SQLException If failed.
-     */
-    @Test
-    public void testWrongCharset_Utf8AsAscii() throws SQLException {
-        checkBulkLoadWithWrongCharset(BULKLOAD_UTF8_CSV_FILE, "UTF-8", "ascii");
-    }
-
-    /**
-     * Verifies that no error is reported and characters are converted improperly when we import
-     * windows-1251 as ASCII.
-     *
-     * @throws SQLException If failed.
-     */
-    @Test
-    public void testWrongCharset_Win1251AsAscii() throws SQLException {
-        checkBulkLoadWithWrongCharset(BULKLOAD_CP1251_CSV_FILE, "windows-1251", "ascii");
-    }
-
-    /**
-     * Checks that no error is reported and characters are converted improperly when we import
-     * file having a different charset than the one specified in the SQL statement.
-     *
-     * @param csvFileName Imported file name.
-     * @param csvCharsetName Imported file charset.
-     * @param stmtCharsetName Charset to specify in the SQL statement.
-     * @throws SQLException If failed.
-     */
-    private void checkBulkLoadWithWrongCharset(String csvFileName, String csvCharsetName, String stmtCharsetName)
-            throws SQLException {
-        int updatesCnt = stmt.executeUpdate(
-                "copy from '" + csvFileName + "' into " + TBL_NAME +
-                        " (_key, age, firstName, lastName)" +
-                        " format csv charset '" + stmtCharsetName + "'");
-
-        assertEquals(2, updatesCnt);
-
-        checkRecodedNationalCacheContents(TBL_NAME, csvCharsetName, stmtCharsetName);
-    }
-
-    /**
      * Checks that bulk load works when we use packet size of 1 byte and thus
      * create multiple packets per COPY.
      *
